Hopefully fix syntax highlighting errors with Scala Map
The errors in some IDEs *might* be caused by an ambiguous type reference. It should not really be ambiguous, but we can make sure it isn't by explicitly specifying the correct type using its FQN.

I haven't found anything else in the `Map`'s bytecode that could confuse the syntax highlighter. It does not override anything, and the type signature for `get` is super-simple.